refactor: renamed it to RasterDefaultLegend
