+++ sites/geohub/src/components/controls/RasterDefaultLegend.svelte
